Check if pids of processes are greater than 0 before killing them

+++ single-node.sh
@@ -413,7 +413,12 @@ stop)
     stop_containers
 
     for pid in $(sudo supervisorctl -c ./config/supervisord.conf pid all); do
-        kill -9 -"${pid}"
+        # If a process is stopped, supervisorctl shows that the pid of the
+        # process is 0. It's not what we need.
+        if [[ "${pid}" > 0 ]]; then
+            info "killing ${pid}"
+            kill -9 -"${pid}"
+        fi
     done
 
     kill -9 "$(sudo supervisorctl -c ./config/supervisord.conf pid)"
